pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "linkedin-agent"
version = "2.0.0"
description = "LinkedIn automation agent suite for outreach, engagement, and networking"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "AI Agent"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "playwright>=1.40.0",
    "python-dotenv>=1.0.0",
    "google-genai>=0.1.0",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "winotify>=1.1.0",
    "fpdf>=1.7.2",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
linkedin-agent = "linkedin_agent.cli:main"

[project.urls]
Homepage = "https://github.com/yourname/linkedin-agent"
Repository = "https://github.com/yourname/linkedin-agent"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
linkedin_agent = ["templates/*.html"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
ignore = ["E501"]
